requirements.txt
requests
slackclient
slackeventsapi
python-dotenv
pymongo
dnspython